Create Readme.md file. Deleted README.md
This is a long description about creating the
Readme.md file. Not enough actions were completed
to warrant this length of a description. Deleted
README.md due to not paying enough attention to
original file name.

This is a brief sentence or two. maybe three
about the problem this commit may or may not solve

Don't forget to link your ref tickets here